Add some defines in package_info
These defines are normally added by "include(${VTK_USE_FILE})" in your
cmakelists, but this only works if you do find_package for VTK in order for the
VTK_USE_FILE variable to be defined in cmake.

Since we are using conan to install VTK we don't want to modify CMakeLists to
add an extra find_package just for the VTK_USE_FILE variable to be defined.
Therefore, it was added to the conan recipe. The caveat is that I'm not sure if
it is valid for other machines.